requirements.txt
asttokens         3.0.0
babel             2.17.0
chardet           5.2.0
colorama          0.4.6
comm              0.2.2
contourpy         1.3.2
customtkinter     5.2.2
cycler            0.12.1
darkdetect        0.8.0
debugpy           1.8.12
decorator         5.1.1
defusedxml        0.7.1
et_xmlfile        2.0.0
executing         2.2.0
fonttools         4.56.0
fpdf              1.7.2
fpdf2             2.8.3
greenlet          3.2.1
ipykernel         6.29.5
ipython           8.32.0
jedi              0.19.2
jupyter_client    8.6.3
jupyter_core      5.7.2
kiwisolver        1.4.8
matplotlib        3.10.1
matplotlib-inline 0.1.7
mysql             0.0.3
mysql-connector   2.2.9
mysqlclient       2.2.7
nest-asyncio      1.6.0
numpy             2.2.5
openpyxl          3.1.5
packaging         24.2
pandas            2.2.3
parso             0.8.4
pillow            11.1.0
pip               25.1
platformdirs      4.3.6
prompt_toolkit    3.0.50
psutil            6.1.1
pure_eval         0.2.3
py                1.11.0
pygame            2.6.1
Pygments          2.19.1
PyMySQL           1.1.1
pyparsing         3.2.3
python-dateutil   2.9.0.post0
pytz              2025.2
pywin32           308
pyzmq             26.2.1
qrcode            8.1
reportlab         4.4.0
six               1.17.0
SQLAlchemy        2.0.40
stack-data        0.6.3
tkcalendar        1.6.1
tksvg             0.7.4
tornado           6.4.2
traitlets         5.14.3
typing_extensions 4.13.2
tzdata            2025.2
wcwidth           0.2.13